+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 groLog
 ============
 first commit
+
+12.07.18 second version
+harvest wet and dry weights not available and show up when harvested
+not much more
+few updates to the functions and a few updates to the sql schema